+++ tests/tests_unit/test_data/test_cache/test_inheritance.py
@@ -2,7 +2,7 @@ from typing import List
 
 from tests.tests_unit.utils import (
     is_cache_file_exists,
-    iterate_data_and_do_checks,
+    iterate_data_and_do_cache_checks,
 )
 
 from th2_data_services.data import Data
@@ -22,7 +22,7 @@ def test_parent_cache_was_created(log_checker, general_data: List[dict]):
     data2 = data1.map(lambda record: {**record, "batch_status": record.get("isBatched")})
     list(data2)  # Just to iterate and create D cache file.
     assert is_cache_file_exists(data)
-    log_checker.cache_file_created(data)
+    # log_checker.cache_file_created(data)
 
     # [2]
     data = Data(general_data, cache=True)
@@ -32,8 +32,8 @@ def test_parent_cache_was_created(log_checker, general_data: List[dict]):
     list(data2)  # Just to iterate and create cache files.
     assert is_cache_file_exists(data)
     assert is_cache_file_exists(data2)
-    log_checker.cache_file_created(data)
-    log_checker.cache_file_created(data2)
+    # log_checker.cache_file_created(data)
+    # log_checker.cache_file_created(data2)
 
 
 def test_data_iterates_parent_cache_file(log_checker, general_data: List[dict]):
@@ -44,8 +44,8 @@ def test_data_iterates_parent_cache_file(log_checker, general_data: List[dict]):
     data1 = data.filter(lambda record: record.get("isBatched"))
     data2 = data1.map(lambda record: {**record, "batch_status": record.get("isBatched")})
     data2.use_cache(True)
-    iterate_data_and_do_checks(data2, log_checker)  # Just to iterate and create cache files.
-    log_checker.cache_file_created(data)
+    iterate_data_and_do_cache_checks(data2, log_checker)  # Just to iterate and create cache files.
+    # log_checker.cache_file_created(data)
 
     data3 = data2.filter(lambda record: record.get("eventType"))
     data4 = data3.map(lambda record: (record, record))
@@ -54,7 +54,7 @@ def test_data_iterates_parent_cache_file(log_checker, general_data: List[dict]):
     data._data_stream = ["D"]
     data2._data_stream = ["D2"]
     assert len(list(data4)) == len(list(data3)) * 2
-    log_checker.used_own_cache_file(data2)
+    # log_checker.used_own_cache_file(data2)
 
 
 def test_cache_linear_inheritance(general_data: List[dict]):